requirements.txt
asgiref==3.2.3
autopep8==1.4.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.1
django-better-filter-widget==0.4.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
mccabe==0.6.1
mysqlclient==1.4.6
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
